refactor: 리팩토링 함수 모듈화 구현
